Added ranking results to edit-fms

+++ src/scripts/page/edit-fms.js
@@ -1,14 +1,16 @@
 /**
  * file:        edit-fms.js
- * description: Allows the user to add and delete FMS OPR and match breakdown results.
+ * description: Allows the user to add and delete FMS ranking, OPR, and match breakdown results.
  * author:      Liam Fruzyna
  * date:        2025-07-25
  */
 
 var new_match_key, new_match_name, negative_match_cb
 var new_copr_key, new_copr_name, negative_copr_cb
+var new_rank_key, new_rank_name, negative_rank_cb
 
-var matches, match_keys, copr_keys
+var matches, rankings
+var match_keys, copr_keys, rank_keys = []
 
 /**
  * Get possible keys.
@@ -16,14 +18,47 @@ var matches, match_keys, copr_keys
 function init_page()
 {
     matches = JSON.parse(localStorage.getItem(`matches-${dal.event_id}`))
-    let breakdown = matches[0].score_breakdown.red
-    match_keys = Object.keys(breakdown).filter(k => 
-        ['boolean', 'number', 'string'].includes(typeof breakdown[k]) &&
-        (k.endsWith('Robot1') || !k.includes('Robot'))
-    ).map(k => k.endsWith('Robot1') ? k.substring(0, k.length - 1) + 'X' : k)
+    if (matches !== null && matches.length > 0)
+    {
+        let breakdown = matches[0].score_breakdown.red
+        match_keys = Object.keys(breakdown).filter(k => 
+            ['boolean', 'number', 'string'].includes(typeof breakdown[k]) &&
+            (k.endsWith('Robot1') || !k.includes('Robot'))
+        ).map(k => k.endsWith('Robot1') ? k.substring(0, k.length - 1) + 'X' : k)
+    }
 
     let coprs = JSON.parse(localStorage.getItem(`coprs-${dal.event_id}`))
     copr_keys = coprs !== null ? Object.keys(coprs) : []
+
+    rankings = JSON.parse(localStorage.getItem(`rankings-${dal.event_id}`))
+    if (rankings !== null && rankings.length > 0)
+    {
+        for (let key of Object.keys(rankings[0]))
+        {
+            let value = rankings[0][key]
+            if (Array.isArray(value))
+            {
+                for (let i in value)
+                {
+                    rank_keys.push(`${key}_${i}`)
+                }
+            }
+            else if (typeof value === 'object')
+            {
+                if (value !== null)
+                {
+                    for (let sub_key of Object.keys(value))
+                    {
+                        rank_keys.push(`${key}_${sub_key}`)
+                    }
+                }
+            }
+            else if (key !== 'team_key')
+            {
+                rank_keys.push(key)
+            }
+        }
+    }
 
     build_buttons()
 }
@@ -33,6 +68,24 @@ function init_page()
  */
 function build_buttons()
 {
+    // build rank page
+    new_rank_key = new WRDropdown('New Key', rank_keys)
+    new_rank_name = new WREntry('New Name', '')
+    negative_rank_cb = new WRCheckbox('Negative?', false)
+    let rank_button = new WRButton('Add Rank Result', build_rank_result)
+
+    let rank_column = new WRColumn('Delete Rank Result')
+    for (let i in cfg.analysis.fms_ranking_results)
+    {
+        let key = cfg.analysis.fms_ranking_results[i].id
+        if (key.startsWith('rank_'))
+        {
+            let name = cfg.get_result_from_key(`fms.${key}`).name
+            rank_column.add_input(new WRButton(name, () => delete_rank_val(i)))
+        }
+    }
+
+    // build C-OPR page
     new_copr_key = new WRDropdown('New Key', copr_keys)
     new_copr_name = new WREntry('New Name', '')
     negative_copr_cb = new WRCheckbox('Negative?', false)
@@ -42,10 +95,14 @@ function build_buttons()
     for (let i in cfg.analysis.fms_ranking_results)
     {
         let key = cfg.analysis.fms_ranking_results[i].id
-        let name = cfg.get_result_from_key(`fms.${key}`).name
-        copr_column.add_input(new WRButton(name, () => delete_copr_val(i)))
+        if (key.startsWith('copr_'))
+        {
+            let name = cfg.get_result_from_key(`fms.${key}`).name
+            copr_column.add_input(new WRButton(name, () => delete_copr_val(i)))
+        }
     }
 
+    // build match page
     new_match_key = new WRDropdown('New Key', match_keys)
     new_match_name = new WREntry('New Name', '')
     negative_match_cb = new WRCheckbox('Negative?', false)
@@ -60,13 +117,50 @@ function build_buttons()
     }
 
     // build template
-    let copr_page = new WRPage('', [new WRColumn('New OPR Result', [new_copr_key, new_copr_name, negative_copr_cb, copr_button]), copr_column])
-    let match_page = new WRPage('', [new WRColumn('New Match Result', [new_match_key, new_match_name, negative_match_cb, match_button]), match_column])
-    preview.replaceChildren(copr_page, match_page)
+    let rank_page = new WRPage('Ranking', [new WRColumn('New Rank Result', [new_rank_key, new_rank_name, negative_rank_cb, rank_button]), rank_column])
+    let copr_page = new WRPage('Component OPR', [new WRColumn('New OPR Result', [new_copr_key, new_copr_name, negative_copr_cb, copr_button]), copr_column])
+    let match_page = new WRPage('Match', [new WRColumn('New Match Result', [new_match_key, new_match_name, negative_match_cb, match_button]), match_column])
+
+    // populate page
+    preview.replaceChildren()
+    if (rank_keys.length)
+    {
+        preview.append(rank_page)
+    }
+    if (copr_keys.length)
+    {
+        preview.append(copr_page)
+    }
+    if (match_keys.length)
+    {
+        preview.append(match_page)
+    }
+}
+
+function determine_type(value)
+{
+    // determine type
+    switch (typeof value)
+    {
+        case 'boolean':
+            return 'boolean'
+        case 'number':
+            return 'int'
+        case 'string':
+            if (['Yes', 'No'].includes(value))
+            {
+                return 'yes_no'
+            }
+            else
+            {
+                return 'state'
+            }
+    }
+    return ''
 }
 
 /**
- * Builds an FMS result based on the selection.
+ * Builds an FMS breakdown result based on the selection.
  */
 function build_match_result()
 {
@@ -81,35 +175,11 @@ function build_match_result()
         return
     }
 
-    // determine type
-    let type = ""
-    switch (typeof value)
-    {
-        case 'boolean':
-            type = 'boolean'
-            break
-        case 'number':
-            type = 'int'
-            break
-        case 'string':
-            if (['Yes', 'No'].includes(value))
-            {
-                type = 'yes_no'
-            }
-            else
-            {
-                type = 'state'
-            }
-            break
-        default:
-            return
-    }
-
     // build id
     let res = {
         id: `bd_${create_id_from_name(id)}`,
         name: name,
-        type: type
+        type: determine_type(value)
     }
 
     // find unique states
@@ -161,7 +231,7 @@ function build_match_result()
 }
 
 /**
- * Builds an FMS result based on the selection.
+ * Builds an FMS C-OPR result based on the selection.
  */
 function build_copr_result()
 {
@@ -187,6 +257,77 @@ function build_copr_result()
 }
 
 /**
+ * Builds an FMS ranking result based on the selection.
+ */
+function build_rank_result()
+{
+    let selected = new_rank_key.element.value
+    let id = `rank_${create_id_from_name(selected)}`
+    let name = new_rank_name.element.value
+
+    if (!name)
+    {
+        alert('Invalid name')
+        return
+    }
+
+    let value = null
+    for (let key of Object.keys(rankings[0]))
+    {
+        let val = rankings[0][key]
+        if (Array.isArray(val))
+        {
+            for (let i in val)
+            {
+                if (selected === `${key}_${i}`)
+                {
+                    value = val[i]
+                    break
+                }
+            }
+            if (value !== null)
+            {
+                break
+            }
+        }
+        else if (typeof val === 'object')
+        {
+            if (val !== null)
+            {
+                for (let sub_key of Object.keys(val))
+                {
+                    if (selected === `${key}_${sub_key}`)
+                    {
+                        value = val[sub_key]
+                        break
+                    }
+                }
+                if (value !== null)
+                {
+                    break
+                }
+            }
+        }
+        else
+        {
+            value = val
+            break
+        }
+    }
+
+    let res = {
+        id: id,
+        name: name,
+        type: determine_type(value),
+        negative: negative_copr_cb.checked
+    }
+
+    cfg.analysis.fms_ranking_results.push(res)
+    cfg.analysis.store_config()
+    build_buttons()
+}
+
+/**
  * Delete the C-OPR at the given index.
  * @param {Number} idx C-OPR index
  */
@@ -198,10 +339,21 @@ function delete_match_val(idx)
 }
 
 /**
- * Delete the match breakdown at the given index.
- * @param {Number} idx Match breakdown index
+ * Delete the C-OPR value at the given index.
+ * @param {Number} idx C-OPR value index
  */
 function delete_copr_val(idx)
+{
+    cfg.analysis.fms_ranking_results.splice(idx, 1)
+    cfg.analysis.store_config()
+    build_buttons()
+}
+
+/**
+ * Delete the ranking value at the given index.
+ * @param {Number} idx Ranking value index
+ */
+function delete_rank_val(idx)
 {
     cfg.analysis.fms_ranking_results.splice(idx, 1)
     cfg.analysis.store_config()
